fix signOut service

+++ src/auth/auth.service.ts
@@ -1,18 +1,13 @@
-import {
-  Injectable,
-  ForbiddenException,
-  Inject,
-  forwardRef,
-} from "@nestjs/common";
-import { UserResponse } from "src/auth/dto/auth.dto";
-import { PrismaService } from "src/prisma/prisma.service";
-import { UserService } from "src/user/user.service";
-import { UtilsService } from "src/utils/utils.service";
-import * as argon2 from "argon2";
-import { ConfigService } from "@nestjs/config";
-import Cryptr from "cryptr";
-import { JwtService } from "@nestjs/jwt";
-import { SessionDto } from "./dto/session.dto";
+import { Injectable, ForbiddenException, Inject, forwardRef } from '@nestjs/common';
+import { UserResponse } from 'src/auth/dto/auth.dto';
+import { PrismaService } from 'src/prisma/prisma.service';
+import { UserService } from 'src/user/user.service';
+import { UtilsService } from 'src/utils/utils.service';
+import * as argon2 from 'argon2';
+import { ConfigService } from '@nestjs/config';
+import Cryptr from 'cryptr';
+import { JwtService } from '@nestjs/jwt';
+import { SessionDto } from './dto/session.dto';
 @Injectable()
 export class AuthService {
   constructor(
@@ -21,40 +16,29 @@ export class AuthService {
     private config: ConfigService,
     private readonly jwt: JwtService,
     private readonly prisma: PrismaService,
-    private readonly utils: UtilsService
+    private readonly utils: UtilsService,
   ) {}
   async refreshTokens(userId: number, refreshToken: string) {
     const user = await this.user.getUserByUserId(userId);
 
     if (!user || !user.refreshToken) {
-      throw new ForbiddenException("Access Denied");
+      throw new ForbiddenException('Access Denied');
     }
 
-    const refreshTokenMatches = await argon2.verify(
-      user.refreshToken,
-      refreshToken
-    );
+    const refreshTokenMatches = await argon2.verify(user.refreshToken, refreshToken);
 
     if (!refreshTokenMatches) {
-      throw new ForbiddenException("Access Denied");
+      throw new ForbiddenException('Access Denied');
     }
 
     const userByPublicId = await this.user.getUserByPublicId(user.publicId);
     const token = await this.getTokens(userByPublicId.userId, user.email);
-    await this.updateRefreshToken(
-      userByPublicId.userId,
-      token.refreshToken,
-      null
-    );
+    await this.updateRefreshToken(userByPublicId.userId, token.refreshToken, null);
 
     return token;
   }
 
-  async updateRefreshToken(
-    userId: number,
-    refreshToken: string,
-    accessToken: string
-  ) {
+  async updateRefreshToken(userId: number, refreshToken: string, accessToken: string) {
     const hashedRefreshToken = await this.hashData(refreshToken);
     if (!accessToken) {
       try {
@@ -96,23 +80,23 @@ export class AuthService {
 
   async getTokens(
     userId: number,
-    email: string
+    email: string,
   ): Promise<{ verifyToken: string; refreshToken: string }> {
     const payload = {
       sub: userId,
       email,
     };
 
-    const secret = this.config.get("JWT_SECRET");
-    const refreshSecret = this.config.get("JWT_REFRESH_SECRET");
+    const secret = this.config.get('JWT_SECRET');
+    const refreshSecret = this.config.get('JWT_REFRESH_SECRET');
 
     const verifyToken = await this.jwt.signAsync(payload, {
-      expiresIn: "1d",
+      expiresIn: '1d',
       secret: secret,
     });
 
     const refreshToken = await this.jwt.signAsync(payload, {
-      expiresIn: "2d",
+      expiresIn: '2d',
       secret: refreshSecret,
     });
     return {
@@ -121,7 +105,7 @@ export class AuthService {
     };
   }
   getPassword(password: string, needCrypt: boolean) {
-    const cryptr = new Cryptr(this.config.get("HASH_SECRET"));
+    const cryptr = new Cryptr(this.config.get('HASH_SECRET'));
     let passwordAfter: string;
     if (needCrypt) {
       passwordAfter = cryptr.encrypt(password);
@@ -132,7 +116,6 @@ export class AuthService {
   }
 
   async createSession(userId: number, session: SessionDto) {
-    console.log(session)
     await this.prisma.session.create({
       data: {
         accessToken: session.accessToken,
@@ -145,6 +128,9 @@ export class AuthService {
 
   async signOut(publicUserId: string) {
     const user = await this.user.getUserByPublicId(publicUserId);
+    if (!user) {
+      throw new ForbiddenException('No user exist');
+    }
     await this.prisma.session.delete({
       where: {
         userId: user.userId,
